cmake refactoring + hunterizations
* add huntergate
* use find_package instead of add_subdirectory fo
   + rabit
   + dmlc-core
   + cub
   + nccl (commented)
* replace global cmake functions with target_ calls to accomodate the migration to pkgconfig targets
* make some options specific to the non R library build
* add XGBOOST_BUILD_SHARED
* add XGBOOST_BUILD_EXAMPLE
* add XGBOOST_BUILD_TEST
* add pkgconfig installation
* generate empty *.cc files required for add_library calls (Xcode generator)

commit 6b659e2f7ea39b701d4dea02af2b564e724d8c45
Author: David Hirvonen <dhirvonen@elucideye.com>
Date:   Sat Jan 6 12:31:40 2018 -0500

    cmake fixes

    * don’t force release build
    * add temporary empty cc file for static lib creation
    * remove set_output_directory

commit 014a754bf961621b3d0762c164abe430d28ea810
Author: David Hirvonen <dhirvonen@elucideye.com>
Date:   Sat Jan 6 12:30:31 2018 -0500

    reset submodule

commit 0cd092c3b9c9fbdc92d0d42b5eff01893504f5e9
Author: David Hirvonen <dhirvonen@elucideye.com>
Date:   Sat Jan 6 12:21:48 2018 -0500

    debug submodule

commit 5920dfee9758b2f622828c7dfed19f50b9863e49
Author: David Hirvonen <dhirvonen@elucideye.com>
Date:   Sat Jan 6 02:40:08 2018 -0500

    hunterize wip

+++ src/tree/updater_gpu_hist.cu
@@ -552,8 +552,8 @@ class GPUHistMaker : public TreeUpdater {
     shards.resize(n_devices);
     row_segments.push_back(0);
     for (int d_idx = 0; d_idx < n_devices; ++d_idx) {
-      bst_uint row_end =
-          std::min(static_cast<size_t>(row_begin + shard_size), info->num_row);
+      bst_uint row_end = // cast to prevent: argument types are: (size_t, uint64_t)
+          std::min(static_cast<size_t>(row_begin + shard_size), static_cast<size_t>(info->num_row));
       row_segments.push_back(row_end);
       row_begin = row_end;
     }
